+++ aspnetcore/src/Repositories/Maps/ResearchDatasetProfile.cs
@@ -24,7 +24,7 @@ public class ResearchDatasetProfile : Profile
             .ForMember(dst => dst.DescriptionEn, opt => opt.MapFrom(src => src.DescriptionEn))
             .ForMember(dst => dst.DatasetCreated, opt => opt.MapFrom(src => src.DatasetCreated))
             .ForMember(dst => dst.Contributors, opt => opt.MapFrom(src => src.FactContributions.Where(fc => fc.DimOrganizationId != -1 || fc.DimNameId != -1)))
-            .ForMember(dst => dst.FieldOfSciences, opt => opt.MapFrom(src => src.DimFieldOfSciences))
+            .ForMember(dst => dst.FieldOfSciences, opt => opt.MapFrom(src => src.FactDimReferencedataFieldOfSciences))
             .ForMember(dst => dst.Languages, opt => opt.MapFrom(src => src.DimReferencedata))
             .ForMember(dst => dst.AccessType, opt => opt.MapFrom(src => src.DimReferencedataAvailabilityNavigation))
             .ForMember(dst => dst.License, opt => opt.MapFrom(src => src.DimReferencedataLicenseNavigation))
@@ -59,11 +59,11 @@ public class ResearchDatasetProfile : Profile
             .ForMember(dst => dst.NameSv, opt => opt.MapFrom(src => src.NameSv))
             .ForMember(dst => dst.NameEn, opt => opt.MapFrom(src => src.NameEn));
 
-        CreateProjection<DimFieldOfScience, FieldOfScience>()
-            .ForMember(dst => dst.FieldId, opt => opt.MapFrom(src => src.FieldId))
-            .ForMember(dst => dst.NameFi, opt => opt.MapFrom(src => src.NameFi))
-            .ForMember(dst => dst.NameSv, opt => opt.MapFrom(src => src.NameSv))
-            .ForMember(dst => dst.NameEn, opt => opt.MapFrom(src => src.NameEn));
+        CreateProjection<FactDimReferencedataFieldOfScience, FieldOfScience>()
+            .ForMember(dst => dst.FieldId, opt => opt.MapFrom(src => src.DimReferencedata.CodeValue))
+            .ForMember(dst => dst.NameFi, opt => opt.MapFrom(src => src.DimReferencedata.NameFi))
+            .ForMember(dst => dst.NameSv, opt => opt.MapFrom(src => src.DimReferencedata.NameSv))
+            .ForMember(dst => dst.NameEn, opt => opt.MapFrom(src => src.DimReferencedata.NameEn));
 
         CreateProjection<DimKeyword, Keyword>()
             .ForMember(dst => dst.Value, opt => opt.MapFrom(src => src.Keyword))